Menus Operation
The workings of how a menu will be displayed and updated. Uses a menu list to store all the menus created, with methods to minipulated those menus. You are alos allowed to choose which menu is going to be currently displayed. Using the DisplayedMenu class. The DisplayedMenu class is connected to the EngineLoop so it is the only menu updated and rendered.